refactor(w3): lift the shared @file list reader out of ApplyTools (ListParams)
housecarl_create's records= takes the same SPEC 5.1 @file convention as
housecarl_apply's ops= / assignments=, and a second copy of the reader is how
the three spellings drift apart - which is the one thing the reader's own doc
comment says it exists to prevent.

Moved verbatim (inline array | "@path" | ["@path"], the strict options that
refuse an undeclared member BY NAME, the STJ position shear, and the
(member, declaring type, correction) vocabulary table), plus create's own rows:
operations -> ops on a record spec, verb -> op on a field op, and the three
members that are not renamed but MEANINGLESS on a create (formid, from,
from_source) each answered in its own words.

No wording changed: apply-guard 63/63 unchanged.
